Added solution - LeetHub

+++ Reverse array in groups - GFG/reverse-array-in-groups.cpp
@@ -13,7 +13,7 @@ public:
     //Function to reverse every sub-array group of size k.
     void reverseInGroups(vector<long long>& arr, int n, int k){
         
-        for(long long i=0;i<n;i+=k)
+        for(int i=0;i<n;i+=k)
         {
             if(i+k<n)
             {
